feat(items): implement SendMailToWinner command and handler, update scheduling logic for email notifications
