+++ backend/src/modules/items/cqrs/commands/implements/send-mail-to-winner.command.ts
@@ -1,0 +1,3 @@
+export class SendMailToWinnerCommand {
+    constructor() { }
+}
