+++ backend/src/modules/items/items.module.ts
@@ -16,8 +16,9 @@ import { GetWinningBidsByUserIdExportPdfQueryHandler } from './cqrs/queries/hand
 import { GetItemByIdExportPdfQueryHandler } from './cqrs/queries/handlers/get-item-by-id-export-pdf.query.handler';
 import { LockItemCommandHandler } from './cqrs/commands/handlers/lock-item.command.handler';
 import { SendMailToWinnerSchedule } from './schedules/send-mail-to-winner.schedule';
+import { SendMailToWinnerCommandHandler } from './cqrs/commands/handlers/send-mail-to-winner.command.handler';
 
-const commandHandlers = [CreateItemCommandHandler, UpdateItemCommandHandler, LockItemCommandHandler];
+const commandHandlers = [CreateItemCommandHandler, UpdateItemCommandHandler, LockItemCommandHandler, SendMailToWinnerCommandHandler];
 const queryHandlers = [
   GetItemByIdQueryHandler,
   GetItemsByOwnerIdHandler,
