+++ backend/src/modules/items/schedules/send-mail-to-winner.schedule.ts
@@ -1,67 +1,15 @@
 import { Injectable } from "@nestjs/common";
+import { CommandBus } from "@nestjs/cqrs";
 import { Cron, CronExpression } from "@nestjs/schedule";
-import { ItemRepository } from "../repository/item.repository";
-import { ItemEntity } from "../entities/item.entity";
-import { QueryBus } from "@nestjs/cqrs";
-import { SendMailWithTemplate } from "modules/mail/cqrs/queries/implements/send-mail-with-template.query";
-import { DataSource } from "typeorm";
+import { SendMailToWinnerCommand } from "../cqrs/commands/implements/send-mail-to-winner.command";
 
 @Injectable()
 export class SendMailToWinnerSchedule {
-    constructor(
-        private readonly itemRepository: ItemRepository,
-        private readonly queryBus: QueryBus,
-        private readonly dataSource: DataSource,
-    ) { }
+    constructor(private readonly commandBus: CommandBus) { }
 
     @Cron(CronExpression.EVERY_MINUTE)
     async handleCron() {
-        const now = new Date();
-
-        const rawItems = await this.itemRepository.findItemsNotNotified(now);
-
-        for (const rawItem of rawItems) {
-            await this.dataSource.transaction(async (manager) => {
-                const item = await manager.getRepository(ItemEntity).findOne({
-                    where: { id: rawItem.id },
-                    lock: { mode: 'pessimistic_write' },
-                    relations: { owner: true, winner: true }
-                });
-
-                if (!item) return;
-
-                await this.queryBus.execute(new SendMailWithTemplate(
-                    item.winner.email,
-                    `Congratulations! You won the auction for "${item.name}"`,
-                    "notify-to-winner.hbs",
-                    { ...this.filterData(item) }
-                ));
-
-                await manager.getRepository(ItemEntity).update(item.id, { isWinnerNotified: true });
-            });
-        }
-
-        console.log(`[${now.toISOString()}] Sent ${rawItems.length} email(s) to winners.`);
+        await this.commandBus.execute(new SendMailToWinnerCommand())
     }
 
-    private filterData(item: ItemEntity) {
-        return {
-            id: item.id,
-            name: item.name,
-            description: item.description,
-            startTime: item.startTime,
-            endTime: item.endTime,
-            startingPrice: item.startingPrice,
-            owner: {
-                id: item.owner.id,
-                email: item.owner.email,
-                name: item.owner.firstName + ' ' + item.owner.lastName
-            },
-            winner: {
-                id: item.winner.id,
-                email: item.winner.email,
-                name: item.winner.firstName + ' ' + item.winner.lastName
-            }
-        }
-    }
 }
